[REFACTORING] userRestControllerTest 코드 수정
